REFACTOR : 자료형 변경 및 hansik 저장 구조 변경

+++ src/main/java/HeoJin/crawling_spring/entity/recipe/CookingOrder.java
@@ -11,6 +11,6 @@ import lombok.NoArgsConstructor;
 @Builder
 @Getter
 public class CookingOrder {
-    private Long step;
+    private Integer step;
     private String instruction;
 }
